Moving flaky tests in the flaky test suite
## What
Moving flaky tests in the flaky test suite.

## Why
We can't make them work reliably on CI server at the  moment and this makes the build break too often.

## How
Moved them in the flaky suite.
