+++ test-cypress/integration/sparql-editor/yasr/table-plugin.spec.js
@@ -2,8 +2,6 @@ import {QueryStubs} from "../../../stubs/yasgui/query-stubs";
 import {SparqlEditorSteps} from "../../../steps/sparql-editor-steps";
 import {YasqeSteps} from "../../../steps/yasgui/yasqe-steps";
 import {YasrSteps} from "../../../steps/yasgui/yasr-steps";
-import {ApplicationSteps} from "../../../steps/application-steps";
-import {TablePluginSteps} from "../../../steps/yasgui/table-plugin-steps";
 
 describe('Yasr Table plugin', () => {
     let repositoryId;
@@ -23,7 +21,7 @@ describe('Yasr Table plugin', () => {
 
     describe('Results formatting', () => {
 
-        it('Should all resource be formatted with short uri when results are of type uri', {
+        it('Should format all resources as short uris when results are of type uri', {
             retries: {
                 runMode: 1,
                 openMode: 0
@@ -37,64 +35,5 @@ describe('Yasr Table plugin', () => {
             YasrSteps.getResultCell(1, 2).contains('rdf:type');
             YasrSteps.getResultCell(4, 3).contains('owl:TransitiveProperty');
         });
-
-        it('Should copy url link be visible when the mouse is over a cell of result table', () => {
-            // When I execute a query which return results and results type is uri.
-            QueryStubs.stubDefaultQueryResponse();
-            YasqeSteps.executeQuery();
-
-            // And I hovered the mouse over a cell of result table.
-            YasrSteps.hoverCell(28, 2);
-
-            // Then I expect copy url link to be visible
-            YasrSteps.getCopyResourceLink(28, 2).should('be.visible');
-        });
-    });
-
-    describe('Copy resource link dialog', () => {
-
-        it('Should be able to copy a link', {
-            retries: {
-                runMode: 1,
-                openMode: 0
-            }
-        }, () => {
-            // When I execute a query which returns results of type is uri.
-            QueryStubs.stubDefaultQueryResponse();
-            YasqeSteps.executeQuery();
-
-            // And copy resource dialog is open.
-            openCopyResourceLinkDialog();
-
-            // Then I expect copy link dialog to be opened.
-            TablePluginSteps.getCopyResourceLinkDialog().should('be.visible');
-
-            // And I expect the input of dialog to have value.
-            TablePluginSteps.getCopyResourceLinkInput().should('have.value', 'http://www.w3.org/1999/02/22-rdf-syntax-ns#type');
-
-            // When I click on cancel button
-            TablePluginSteps.clickCopyLinkDialogCancelButton();
-
-            // Then I expect copy link dialog to be closed.
-            TablePluginSteps.getCopyResourceLinkDialog().should('not.exist');
-
-            // And copy resource dialog is open.
-            openCopyResourceLinkDialog();
-
-            // And click on copy button
-            TablePluginSteps.clickCopyLinkDialogCopyButton();
-
-            // Then I expect copy link dialog to be closed.
-            TablePluginSteps.getCopyResourceLinkDialog().should('not.exist');
-
-            // And expect success message to be displayed.
-            ApplicationSteps.getSuccessNotifications().contains('URL copied successfully to clipboard.');
-
-        });
     });
 });
-
-function openCopyResourceLinkDialog(rowNumber = 28, cellNumber = 2) {
-    YasrSteps.getCopyResourceLink(rowNumber, cellNumber).should('be.visible');
-    YasrSteps.copyResourceLink(rowNumber, cellNumber);
-}
